+++ src/mailman/database/schema/mm_20121015000000.py
@@ -96,6 +96,7 @@ def upgrade_postgres(database, store, version, module_path):
     store.execute('ALTER TABLE mailinglist DROP COLUMN send_reminders;')
     store.execute('ALTER TABLE mailinglist DROP COLUMN subscribe_policy;')
     store.execute('ALTER TABLE mailinglist DROP COLUMN unsubscribe_policy;')
+    store.execute('ALTER TABLE mailinglist DROP COLUMN private_roster;')
     store.execute(
         'ALTER TABLE mailinglist DROP COLUMN subscribe_auto_approval;')
     # Record the migration in the version table.