+++ src/mailman/database/tests/test_migrations.py
@@ -380,7 +380,7 @@ class TestMigration20121015Schema(MigrationTestBase):
                               (),
                               ('new_member_options', 'send_reminders',
                                'subscribe_policy', 'unsubscribe_policy',
-                               'subscribe_auto_approval'),
+                               'subscribe_auto_approval', 'private_roster'),
                               )
 
     def test_post_upgrade_column_migrations(self):
@@ -394,7 +394,7 @@ class TestMigration20121015Schema(MigrationTestBase):
                                '20121015000000'],
                               ('new_member_options', 'send_reminders',
                                'subscribe_policy', 'unsubscribe_policy',
-                               'subscribe_auto_approval'),
+                               'subscribe_auto_approval', 'private_roster'),
                               ())